Skin Widget Recommended TV Shows
I've edited the includes_shelf.xml so it now chooses recommended episodes instead of recently added episodes.
Pecinko, this is not (yet?) a toggle feature, it's a dirty hack ;). 
If you are interested in supporting this in the master branch you'll probably need to combine the original file and my file.